requirements/test.txt
coverage==4.2
django-dynamic-fixture==1.9.0
flake8==3.0.4
mock==2.0.0
pytest==3.0.4
pytest-django==3.1.1
pytest-pythonpath==0.7.1
pytest-sugar==0.8.0
requests_mock==1.3.0
Faker==0.7.3
vcrpy==1.11.0
pytest-mock==1.6.3